feat(api): add support for conditional component.
This commit allows defining the conditions that need to be fulfilled for
a component to be eligible for use in the application.

Conditions can be defined either using new annotations or ComponentDescriptorModifiers.

This commit introduce a set of new classes :

  - add new annotation classes :  ConditionalOn and Conditionals
  - add new interfaces/classes : Condition, Conditions, ConditionContext

This commit also refactor the built-in StreamsLifecycleInterceptor interface
implementations.

Breaking changes :

  - remove method AzkarraContext#setComponentFactory
  - StreamsLifecycleContext some methods have been rename
  - AutoCreateTopicsInterceptor#setConfigs renamed to setTopicConfigs
  - ConflictingBeanDefinitionException renamed to ConflictingComponentDefinitionException
  - ComponentFactory#getComponent/getAllComponents renamed to
getComponentProvider and getAllComponentProvides