feat(api): add the capability to eagerly initialize a component
By default, Azkarra lazily instantiates and configures a component when it is first requested.
Generally, this is the behavior desired by users because the contextual configuration of a
component cannot be resolved at context startup.

This commit add capability for singleton component to be eagerly created
and configured at context startup.

 - add new annotation Eager
 - add new method ComponentDescriptorModifiers#asEager